+++ 排序算法/插入排序.c
@@ -3,12 +3,12 @@
 // 其实这种方法比较低效，它先从头开始找位置，然后又从后往前每个往后移
 void InsertSort(size_t len, int arr[static restrict len])
 {
-    for (int i = 1; i < len; i++)   // a[0]认为已经排序好了，之后用DP/数学归纳的思想
-        for (int j = 0; j < i; j++) // 每次取arr[i]插入前面的数组
+    for (size_t i = 1; i < len; i++)   // a[0]认为已经排序好了，之后用DP/数学归纳的思想
+        for (size_t j = 0; j < i; j++) // 每次取arr[i]插入前面的数组
             if (arr[j] > arr[i])    // arr[j]更小时就什么也不做
             {
                 int t = arr[i];
-                for (int k = i; k > j; k--) // 往后移，可用memmove替代
+                for (size_t k = i; k > j; k--) // 往后移，可用memmove替代
                     arr[k] = arr[k - 1];
                 arr[j] = t;
             }
